feat(refactor): merge webpack configuration
Taken from
 - https://github.com/facebook/create-react-app/commit/30ee52cf3b2cbb6ac70999c02b1196bcaba8d4ca

- https://github.com/facebook/create-react-app/pull/5722

But adjusted to match our needs
